work in progress Extracted GlFontDrawingPrimitive
