+++ QuickFont/QFont.cs
@@ -8,7 +8,10 @@ using OpenTK.Graphics.OpenGL4;
 
 namespace QuickFont
 {
-    public class QFont
+    /// <summary>
+    /// This class represents a 'Drawing' on screen made off of Font renderings. eg. GlFontDrawingPrimitive
+    /// </summary>
+    public class QFontDrawing
     {
         //private QFontRenderOptions options = new QFontRenderOptions();
         private const string fragShaderSource = @"#version 130
@@ -48,19 +51,7 @@ void main(void)
         public QVertexArrayObject[] VertexArrayObjects = new QVertexArrayObject[0];
         private SharedState _instanceSharedState;
 
-        private Vector3 _printOffset;
         private Matrix4 _projectionMatrix;
-
-        public Vector3 PrintOffset
-        {
-            get { return _printOffset; }
-            set
-            {
-                _printOffset = value;
-                if (GlFont.FontData.dropShadow != null)
-                    GlFont.FontData.dropShadow.PrintOffset = value;
-            }
-        }
 
         public QFontRenderOptions Options
         {
@@ -97,25 +88,10 @@ void main(void)
             set { _projectionMatrix = value; }
         }
 
-        public float LineSpacing
-        {
-            get { return (float) Math.Ceiling(GlFont.FontData.maxGlyphHeight*Options.LineSpacing); }
-        }
 
-        public bool IsMonospacingActive
+        public GlFontDrawingPimitive GlFontDrawingPimitive
         {
-            get { return GlFont.FontData.IsMonospacingActive(Options); }
-        }
-
-
-        public float MonoSpaceWidth
-        {
-            get { return GlFont.FontData.GetMonoSpaceWidth(Options); }
-        }
-
-        public GlFont GlFont
-        {
-            get { return _glFont; }
+            get { return _glFontDrawingPimitive; }
         }
 
         #region Constructors and font builders
@@ -126,7 +102,7 @@ void main(void)
         /// <param name="fontData"></param>
         internal QFont(QFontData fontData)
         {
-            _glFont = new GlFont(fontData);
+            _glFontDrawingPimitive = new GlFontDrawingPimitive(new GlFont(fontData));
         }
 
         /// <summary>
@@ -147,7 +123,8 @@ void main(void)
 
             using (Font font = QuickFont.GlFont.GetFont(fontPath, size, style, config == null ? 1 : config.SuperSampleLevels, fontScale))
             {
-                _glFont = new GlFont(font, config);
+                _glFontDrawingPimitive = new GlFontDrawingPimitive(this);
+                GlFontDrawingPimitive.GlFont = new GlFont(font, config);
             }
 
             if (transToVp != null)
@@ -163,15 +140,12 @@ void main(void)
         /// <param name="proj"></param>
         public QFont(string qfontPath, QFontConfiguration loaderConfig, float downSampleFactor = 1.0f, Matrix4 projectionMatrix = default(Matrix4))
         {
-            _glFont = new GlFont(this);
+            _glFontDrawingPimitive = new GlFontDrawingPimitive(new GlFont(qfontPath, loaderConfig, downSampleFactor, projectionMatrix), this);
             _projectionMatrix = projectionMatrix;
             Viewport? transToVp = null;
             float fontScale = 1f;
             if (loaderConfig.TransformToCurrentOrthogProjection)
                 transToVp = OrthogonalTransform(out fontScale);
-
-            GlFont.InitialiseQFont(null, new QFontBuilderConfiguration(loaderConfig), Builder.LoadQFontDataFromFile(qfontPath, downSampleFactor*fontScale, loaderConfig));
-            ViewportHelper.CurrentViewport.ToString();
 
             if (transToVp != null)
                 Options.TransformToViewport = transToVp;
@@ -263,8 +237,8 @@ void main(void)
             else
             {
                 _instanceSharedState = state;
-                if (GlFont.FontData.dropShadow != null)
-                    GlFont.FontData.dropShadow._instanceSharedState = state;
+                if (GlFontDrawingPimitive.GlFont.FontData.dropShadow != null)
+                    GlFontDrawingPimitive.GlFont.FontData.dropShadow._instanceSharedState = state;
             }
         }
 
@@ -321,788 +295,6 @@ void main(void)
             }
         }
 
-        private void RenderDropShadow(float x, float y, char c, QFontGlyph nonShadowGlyph)
-        {
-            //note can cast drop shadow offset to int, but then you can't move the shadow smoothly...
-            if (GlFont.FontData.dropShadow != null && Options.DropShadowActive)
-            {
-                GlFont.FontData.dropShadow.RenderGlyph(
-                    x + (GlFont.FontData.meanGlyphWidth*Options.DropShadowOffset.X + nonShadowGlyph.rect.Width*0.5f),
-                    y +
-                    (GlFont.FontData.meanGlyphWidth*Options.DropShadowOffset.Y + nonShadowGlyph.rect.Height*0.5f +
-                     nonShadowGlyph.yOffset), c);
-            }
-        }
-
-        public void RenderGlyph(float x, float y, char c)
-        {
-            QFontGlyph glyph = GlFont.FontData.CharSetMapping[c];
-
-            //note: it's not immediately obvious, but this combined with the paramteters to 
-            //RenderGlyph for the shadow mean that we render the shadow centrally (despite it being a different size)
-            //under the glyph
-            if (GlFont.FontData.isDropShadow)
-            {
-                x -= (int) (glyph.rect.Width*0.5f);
-                y -= (int) (glyph.rect.Height*0.5f + glyph.yOffset);
-            }
-            else
-                RenderDropShadow(x, y, c, glyph);
-
-            y = -y;
-
-            TexturePage sheet = GlFont.FontData.Pages[glyph.page];
-
-            float tx1 = (float) (glyph.rect.X)/sheet.Width;
-            float ty1 = (float) (glyph.rect.Y)/sheet.Height;
-            float tx2 = (float) (glyph.rect.X + glyph.rect.Width)/sheet.Width;
-            float ty2 = (float) (glyph.rect.Y + glyph.rect.Height)/sheet.Height;
-
-            var tv1 = new Vector2(tx1, ty1);
-            var tv2 = new Vector2(tx1, ty2);
-            var tv3 = new Vector2(tx2, ty2);
-            var tv4 = new Vector2(tx2, ty1);
-
-            Vector3 v1 = PrintOffset + new Vector3(x, y - glyph.yOffset, 0);
-            Vector3 v2 = PrintOffset + new Vector3(x, y - glyph.yOffset - glyph.rect.Height, 0);
-            Vector3 v3 = PrintOffset + new Vector3(x + glyph.rect.Width, y - glyph.yOffset - glyph.rect.Height, 0);
-            Vector3 v4 = PrintOffset + new Vector3(x + glyph.rect.Width, y - glyph.yOffset, 0);
-
-            Color color;
-            if (GlFont.FontData.isDropShadow)
-                color = Options.DropShadowColour;
-            else
-                color = Options.Colour;
-
-            var normal = new Vector3(0, 0, -1);
-
-            int argb = Helper.ToRgba(color);
-
-            QVertexArrayObject vbo = VertexArrayObjects[glyph.page];
-
-            Vector4 colour = Helper.ToVector4(color);
-
-            vbo.AddVertex(v1, tv1, colour);
-            vbo.AddVertex(v2, tv2, colour);
-            vbo.AddVertex(v3, tv3, colour);
-
-            vbo.AddVertex(v1, tv1, colour);
-            vbo.AddVertex(v3, tv3, colour);
-            vbo.AddVertex(v4, tv4, colour);
-        }
-
-        private float MeasureNextlineLength(string text)
-        {
-            float xOffset = 0;
-
-            for (int i = 0; i < text.Length; i++)
-            {
-                char c = text[i];
-
-                if (c == '\r' || c == '\n')
-                {
-                    break;
-                }
-
-
-                if (IsMonospacingActive)
-                {
-                    xOffset += MonoSpaceWidth;
-                }
-                else
-                {
-                    //space
-                    if (c == ' ')
-                    {
-                        xOffset += (float) Math.Ceiling(GlFont.FontData.meanGlyphWidth*Options.WordSpacing);
-                    }
-                        //normal character
-                    else if (GlFont.FontData.CharSetMapping.ContainsKey(c))
-                    {
-                        QFontGlyph glyph = GlFont.FontData.CharSetMapping[c];
-                        xOffset +=
-                            (float)
-                                Math.Ceiling(glyph.rect.Width + GlFont.FontData.meanGlyphWidth*Options.CharacterSpacing + GlFont.FontData.GetKerningPairCorrection(i, text, null));
-                    }
-                }
-            }
-            return xOffset;
-        }
-
-        private Vector2 TransformPositionToViewport(Vector2 input)
-        {
-            Viewport? v2 = Options.TransformToViewport;
-            if (v2 == null)
-            {
-                return input;
-            }
-            Viewport? v1 = ViewportHelper.CurrentViewport;
-
-            float X, Y;
-
-            Debug.Assert(v1 != null, "v1 != null");
-            X = (input.X - v2.Value.X)*(v1.Value.Width/v2.Value.Width);
-            Y = (input.Y - v2.Value.Y)*(v1.Value.Height/v2.Value.Height);
-
-            return new Vector2(X, Y);
-        }
-
-        private float TransformWidthToViewport(float input)
-        {
-            Viewport? v2 = Options.TransformToViewport;
-            if (v2 == null)
-            {
-                return input;
-            }
-            Viewport? v1 = ViewportHelper.CurrentViewport;
-
-            Debug.Assert(v1 != null, "v1 != null");
-            return input*(v1.Value.Width/v2.Value.Width);
-        }
-
-        private SizeF TransformMeasureFromViewport(SizeF input)
-        {
-            Viewport? v2 = Options.TransformToViewport;
-            if (v2 == null)
-            {
-                return input;
-            }
-            Viewport? v1 = ViewportHelper.CurrentViewport;
-
-            float X, Y;
-
-            Debug.Assert(v1 != null, "v1 != null");
-            X = input.Width*(v2.Value.Width/v1.Value.Width);
-            Y = input.Height*(v2.Value.Height/v1.Value.Height);
-
-            return new SizeF(X, Y);
-        }
-
-        private Vector2 LockToPixel(Vector2 input)
-        {
-            if (Options.LockToPixel)
-            {
-                float r = Options.LockToPixelRatio;
-                return new Vector2((1 - r)*input.X + r*((int) Math.Round(input.X)),
-                    (1 - r)*input.Y + r*((int) Math.Round(input.Y)));
-            }
-            return input;
-        }
-
-        private Vector3 TransformToViewport(Vector3 input)
-        {
-            return new Vector3(LockToPixel(TransformPositionToViewport(input.Xy))) {Z = input.Z};
-        }
-
-        public SizeF Print(string text, Vector3 position, SizeF maxSize, QFontAlignment alignment)
-        {
-            ProcessedText processedText = ProcessText(text, maxSize, alignment);
-            return Print(processedText, TransformToViewport(position));
-        }
-
-        public SizeF Print(string text, Vector3 position, SizeF maxSize, QFontAlignment alignment, Color colour)
-        {
-            ProcessedText processedText = ProcessText(text, maxSize, alignment);
-            return Print(processedText, TransformToViewport(position), colour);
-        }
-
-        public SizeF Print(ProcessedText processedText, Vector3 position)
-        {
-            PrintOffset = TransformToViewport(position);
-            return PrintOrMeasure(processedText, false);
-        }
-
-        public SizeF Print(ProcessedText processedText, Vector3 position, Color colour)
-        {
-            Options.Colour = colour;
-            PrintOffset = TransformToViewport(position);
-            return PrintOrMeasure(processedText, false);
-        }
-
-        public SizeF Print(string text, Vector3 position, QFontAlignment alignment)
-        {
-            PrintOffset = TransformToViewport(position);
-            return PrintOrMeasure(text, alignment, false);
-        }
-
-        public SizeF Print(string text, Vector3 position, QFontAlignment alignment, Color color)
-        {
-            Options.Colour = color;
-            PrintOffset = TransformToViewport(position);
-            return PrintOrMeasure(text, alignment, false);
-        }
-
-        public SizeF Measure(string text, QFontAlignment alignment = QFontAlignment.Left)
-        {
-            return TransformMeasureFromViewport(PrintOrMeasure(text, alignment, true));
-        }
-
-        public SizeF Measure(string text, float maxWidth, QFontAlignment alignment)
-        {
-            return Measure(text, new SizeF(maxWidth, -1), alignment);
-        }
-
-        /// <summary>
-        ///     Measures the actual width and height of the block of text.
-        /// </summary>
-        /// <param name="text"></param>
-        /// <param name="bounds"></param>
-        /// <param name="alignment"></param>
-        /// <returns></returns>
-        public SizeF Measure(string text, SizeF maxSize, QFontAlignment alignment)
-        {
-            ProcessedText processedText = ProcessText(text, maxSize, alignment);
-            return Measure(processedText);
-        }
-
-        /// <summary>
-        ///     Measures the actual width and height of the block of text
-        /// </summary>
-        /// <param name="processedText"></param>
-        /// <returns></returns>
-        public SizeF Measure(ProcessedText processedText)
-        {
-            return TransformMeasureFromViewport(PrintOrMeasure(processedText, true));
-        }
-
-        private SizeF PrintOrMeasure(string text, QFontAlignment alignment, bool measureOnly)
-        {
-            float maxWidth = 0f;
-            float xOffset = 0f;
-            float yOffset = 0f;
-
-            //make sure fontdata font's options are synced with the actual options
-            if (GlFont.FontData.dropShadow != null && GlFont.FontData.dropShadow.Options != Options)
-            {
-                GlFont.FontData.dropShadow.Options = Options;
-            }
-
-            float maxXpos = float.MinValue;
-            float minXPos = float.MaxValue;
-
-            text = text.Replace("\r\n", "\r");
-
-            if (alignment == QFontAlignment.Right)
-                xOffset -= MeasureNextlineLength(text);
-            else if (alignment == QFontAlignment.Centre)
-                xOffset -= (int) (0.5f*MeasureNextlineLength(text));
-
-            for (int i = 0; i < text.Length; i++)
-            {
-                char c = text[i];
-
-                //newline
-                if (c == '\r' || c == '\n')
-                {
-                    yOffset += LineSpacing;
-                    xOffset = 0f;
-
-                    if (alignment == QFontAlignment.Right)
-                        xOffset -= MeasureNextlineLength(text.Substring(i + 1));
-                    else if (alignment == QFontAlignment.Centre)
-                        xOffset -= (int) (0.5f*MeasureNextlineLength(text.Substring(i + 1)));
-                }
-                else
-                {
-                    minXPos = Math.Min(xOffset, minXPos);
-
-                    //normal character
-                    if (c != ' ' && GlFont.FontData.CharSetMapping.ContainsKey(c))
-                    {
-                        QFontGlyph glyph = GlFont.FontData.CharSetMapping[c];
-                        if (!measureOnly)
-                            RenderGlyph(xOffset, yOffset, c);
-                    }
-
-                    if (IsMonospacingActive)
-                        xOffset += MonoSpaceWidth;
-                    else
-                    {
-                        if (c == ' ')
-                            xOffset += (float) Math.Ceiling(GlFont.FontData.meanGlyphWidth*Options.WordSpacing);
-                            //normal character
-                        else if (GlFont.FontData.CharSetMapping.ContainsKey(c))
-                        {
-                            QFontGlyph glyph = GlFont.FontData.CharSetMapping[c];
-                            xOffset +=
-                                (float)
-                                    Math.Ceiling(glyph.rect.Width + GlFont.FontData.meanGlyphWidth*Options.CharacterSpacing + GlFont.FontData.GetKerningPairCorrection(i, text, null));
-                        }
-                    }
-
-                    maxXpos = Math.Max(xOffset, maxXpos);
-                }
-            }
-
-            if (minXPos != float.MaxValue)
-                maxWidth = maxXpos - minXPos;
-
-            return new SizeF(maxWidth, yOffset + LineSpacing);
-        }
-
-        private SizeF PrintOrMeasure(ProcessedText processedText, bool measureOnly)
-        {
-            // init values we'll return
-            float maxMeasuredWidth = 0f;
-
-            float xPos = 0f;
-            float yPos = 0f;
-
-            float xOffset = xPos;
-            float yOffset = yPos;
-
-            //make sure fontdata font's options are synced with the actual options
-            if (GlFont.FontData.dropShadow != null && GlFont.FontData.dropShadow.Options != Options)
-            {
-                GlFont.FontData.dropShadow.Options = Options;
-            }
-
-            float maxWidth = processedText.maxSize.Width;
-            QFontAlignment alignment = processedText.alignment;
-
-
-            //TODO - use these instead of translate when rendering by position (at some point)
-
-            TextNodeList nodeList = processedText.textNodeList;
-            for (TextNode node = nodeList.Head; node != null; node = node.Next)
-                node.LengthTweak = 0f; //reset tweaks
-
-
-            if (alignment == QFontAlignment.Right)
-                xOffset -= (float) Math.Ceiling(TextNodeLineLength(nodeList.Head, maxWidth) - maxWidth);
-            else if (alignment == QFontAlignment.Centre)
-                xOffset -= (float) Math.Ceiling(0.5f*TextNodeLineLength(nodeList.Head, maxWidth));
-            else if (alignment == QFontAlignment.Justify)
-                JustifyLine(nodeList.Head, maxWidth);
-
-
-            bool atLeastOneNodeCosumedOnLine = false;
-            float length = 0f;
-            for (TextNode node = nodeList.Head; node != null; node = node.Next)
-            {
-                bool newLine = false;
-
-                if (node.Type == TextNodeType.LineBreak)
-                {
-                    newLine = true;
-                }
-                else
-                {
-                    if (Options.WordWrap && SkipTrailingSpace(node, length, maxWidth) && atLeastOneNodeCosumedOnLine)
-                    {
-                        newLine = true;
-                    }
-                    else if (length + node.ModifiedLength <= maxWidth || !atLeastOneNodeCosumedOnLine)
-                    {
-                        atLeastOneNodeCosumedOnLine = true;
-
-                        if (!measureOnly)
-                            RenderWord(xOffset + length, yOffset, node);
-                        length += node.ModifiedLength;
-
-                        maxMeasuredWidth = Math.Max(length, maxMeasuredWidth);
-                    }
-                    else if (Options.WordWrap)
-                    {
-                        newLine = true;
-                        if (node.Previous != null)
-                            node = node.Previous;
-                    }
-                    else
-                        continue; // continue so we still read line breaks even if reached max width
-                }
-
-                if (newLine)
-                {
-                    if (processedText.maxSize.Height > 0 &&
-                        yOffset + LineSpacing - yPos >= processedText.maxSize.Height)
-                        break;
-
-                    yOffset += LineSpacing;
-                    xOffset = xPos;
-                    length = 0f;
-                    atLeastOneNodeCosumedOnLine = false;
-
-                    if (node.Next != null)
-                    {
-                        if (alignment == QFontAlignment.Right)
-                            xOffset -= (float) Math.Ceiling(TextNodeLineLength(node.Next, maxWidth) - maxWidth);
-                        else if (alignment == QFontAlignment.Centre)
-                            xOffset -= (float) Math.Ceiling(0.5f*TextNodeLineLength(node.Next, maxWidth));
-                        else if (alignment == QFontAlignment.Justify)
-                            JustifyLine(node.Next, maxWidth);
-                    }
-                }
-            }
-
-            return new SizeF(maxMeasuredWidth, yOffset + LineSpacing - yPos);
-        }
-
-        private void RenderWord(float x, float y, TextNode node)
-        {
-            if (node.Type != TextNodeType.Word)
-                return;
-
-            int charGaps = node.Text.Length - 1;
-            bool isCrumbleWord = CrumbledWord(node);
-            if (isCrumbleWord)
-                charGaps++;
-
-            int pixelsPerGap = 0;
-            int leftOverPixels = 0;
-
-            if (charGaps != 0)
-            {
-                pixelsPerGap = (int) node.LengthTweak/charGaps;
-                leftOverPixels = (int) node.LengthTweak - pixelsPerGap*charGaps;
-            }
-
-            for (int i = 0; i < node.Text.Length; i++)
-            {
-                char c = node.Text[i];
-                if (GlFont.FontData.CharSetMapping.ContainsKey(c))
-                {
-                    QFontGlyph glyph = GlFont.FontData.CharSetMapping[c];
-
-                    RenderGlyph(x, y, c);
-
-
-                    if (IsMonospacingActive)
-                        x += MonoSpaceWidth;
-                    else
-                        x +=
-                            (int)
-                                Math.Ceiling(glyph.rect.Width + GlFont.FontData.meanGlyphWidth*Options.CharacterSpacing + GlFont.FontData.GetKerningPairCorrection(i, node.Text, node));
-
-                    x += pixelsPerGap;
-                    if (leftOverPixels > 0)
-                    {
-                        x += 1.0f;
-                        leftOverPixels--;
-                    }
-                    else if (leftOverPixels < 0)
-                    {
-                        x -= 1.0f;
-                        leftOverPixels++;
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        ///     Computes the length of the next line, and whether the line is valid for
-        ///     justification.
-        /// </summary>
-        /// <param name="node"></param>
-        /// <param name="maxLength"></param>
-        /// <param name="justifable"></param>
-        /// <returns></returns>
-        private float TextNodeLineLength(TextNode node, float maxLength)
-        {
-            if (node == null)
-                return 0;
-
-            bool atLeastOneNodeCosumedOnLine = false;
-            float length = 0;
-            for (; node != null; node = node.Next)
-            {
-                if (node.Type == TextNodeType.LineBreak)
-                    break;
-
-                if (SkipTrailingSpace(node, length, maxLength) && atLeastOneNodeCosumedOnLine)
-                    break;
-
-                if (length + node.Length <= maxLength || !atLeastOneNodeCosumedOnLine)
-                {
-                    atLeastOneNodeCosumedOnLine = true;
-                    length += node.Length;
-                }
-                else
-                {
-                    break;
-                }
-            }
-            return length;
-        }
-
-
-        private bool CrumbledWord(TextNode node)
-        {
-            return (node.Type == TextNodeType.Word && node.Next != null && node.Next.Type == TextNodeType.Word);
-        }
-
-        /// <summary>
-        ///     Computes the length of the next line, and whether the line is valid for
-        ///     justification.
-        /// </summary>
-        private void JustifyLine(TextNode node, float targetLength)
-        {
-            bool justifiable = false;
-
-            if (node == null)
-                return;
-
-            TextNode headNode = node; //keep track of the head node
-
-
-            //start by finding the length of the block of text that we know will actually fit:
-
-            int charGaps = 0;
-            int spaceGaps = 0;
-
-            bool atLeastOneNodeCosumedOnLine = false;
-            float length = 0;
-            TextNode expandEndNode = node; //the node at the end of the smaller list (before adding additional word)
-            for (; node != null; node = node.Next)
-            {
-                if (node.Type == TextNodeType.LineBreak)
-                    break;
-
-                if (SkipTrailingSpace(node, length, targetLength) && atLeastOneNodeCosumedOnLine)
-                {
-                    justifiable = true;
-                    break;
-                }
-
-                if (length + node.Length < targetLength || !atLeastOneNodeCosumedOnLine)
-                {
-                    expandEndNode = node;
-
-                    if (node.Type == TextNodeType.Space)
-                        spaceGaps++;
-
-                    if (node.Type == TextNodeType.Word)
-                    {
-                        charGaps += (node.Text.Length - 1);
-
-                        //word was part of a crumbled word, so there's an extra char cap between the two words
-                        if (CrumbledWord(node))
-                            charGaps++;
-                    }
-
-                    atLeastOneNodeCosumedOnLine = true;
-                    length += node.Length;
-                }
-                else
-                {
-                    justifiable = true;
-                    break;
-                }
-            }
-
-            //now we check how much additional length is added by adding an additional word to the line
-            float extraLength = 0f;
-            int extraSpaceGaps = 0;
-            int extraCharGaps = 0;
-            bool contractPossible = false;
-            TextNode contractEndNode = null;
-            for (node = expandEndNode.Next; node != null; node = node.Next)
-            {
-                if (node.Type == TextNodeType.LineBreak)
-                    break;
-
-                if (node.Type == TextNodeType.Space)
-                {
-                    extraLength += node.Length;
-                    extraSpaceGaps++;
-                }
-                else if (node.Type == TextNodeType.Word)
-                {
-                    contractEndNode = node;
-                    contractPossible = true;
-                    extraLength += node.Length;
-                    extraCharGaps += (node.Text.Length - 1);
-                    break;
-                }
-            }
-
-            if (justifiable)
-            {
-                //last part of this condition is to ensure that the full contraction is possible (it is all or nothing with contractions, since it looks really bad if we don't manage the full)
-                bool contract = contractPossible &&
-                                (extraLength + length - targetLength)*Options.JustifyContractionPenalty <
-                                (targetLength - length) &&
-                                ((targetLength - (length + extraLength + 1))/targetLength > -Options.JustifyCapContract);
-
-                if ((!contract && length < targetLength) || (contract && length + extraLength > targetLength))
-                    //calculate padding pixels per word and char
-                {
-                    if (contract)
-                    {
-                        length += extraLength + 1;
-                        charGaps += extraCharGaps;
-                        spaceGaps += extraSpaceGaps;
-                    }
-
-
-                    var totalPixels = (int) (targetLength - length);
-                    //the total number of pixels that need to be added to line to justify it
-                    int spacePixels = 0; //number of pixels to spread out amongst spaces
-                    int charPixels = 0; //number of pixels to spread out amongst char gaps
-
-
-                    if (contract)
-                    {
-                        if (totalPixels/targetLength < -Options.JustifyCapContract)
-                            totalPixels = (int) (-Options.JustifyCapContract*targetLength);
-                    }
-                    else
-                    {
-                        if (totalPixels/targetLength > Options.JustifyCapExpand)
-                            totalPixels = (int) (Options.JustifyCapExpand*targetLength);
-                    }
-
-                    //work out how to spread pixles between character gaps and word spaces
-                    if (charGaps == 0)
-                    {
-                        spacePixels = totalPixels;
-                    }
-                    else if (spaceGaps == 0)
-                    {
-                        charPixels = totalPixels;
-                    }
-                    else
-                    {
-                        if (contract)
-                            charPixels =
-                                (int) (totalPixels*Options.JustifyCharacterWeightForContract*charGaps/spaceGaps);
-                        else
-                            charPixels = (int) (totalPixels*Options.JustifyCharacterWeightForExpand*charGaps/spaceGaps);
-
-
-                        if ((!contract && charPixels > totalPixels) ||
-                            (contract && charPixels < totalPixels))
-                            charPixels = totalPixels;
-
-                        spacePixels = totalPixels - charPixels;
-                    }
-
-
-                    int pixelsPerChar = 0; //minimum number of pixels to add per char
-                    int leftOverCharPixels = 0; //number of pixels remaining to only add for some chars
-
-                    if (charGaps != 0)
-                    {
-                        pixelsPerChar = charPixels/charGaps;
-                        leftOverCharPixels = charPixels - pixelsPerChar*charGaps;
-                    }
-
-                    int pixelsPerSpace = 0; //minimum number of pixels to add per space
-                    int leftOverSpacePixels = 0; //number of pixels remaining to only add for some spaces
-
-                    if (spaceGaps != 0)
-                    {
-                        pixelsPerSpace = spacePixels/spaceGaps;
-                        leftOverSpacePixels = spacePixels - pixelsPerSpace*spaceGaps;
-                    }
-
-                    //now actually iterate over all nodes and set tweaked length
-                    for (node = headNode; node != null; node = node.Next)
-                    {
-                        if (node.Type == TextNodeType.Space)
-                        {
-                            node.LengthTweak = pixelsPerSpace;
-                            if (leftOverSpacePixels > 0)
-                            {
-                                node.LengthTweak += 1;
-                                leftOverSpacePixels--;
-                            }
-                            else if (leftOverSpacePixels < 0)
-                            {
-                                node.LengthTweak -= 1;
-                                leftOverSpacePixels++;
-                            }
-                        }
-                        else if (node.Type == TextNodeType.Word)
-                        {
-                            int cGaps = (node.Text.Length - 1);
-                            if (CrumbledWord(node))
-                                cGaps++;
-
-                            node.LengthTweak = cGaps*pixelsPerChar;
-
-
-                            if (leftOverCharPixels >= cGaps)
-                            {
-                                node.LengthTweak += cGaps;
-                                leftOverCharPixels -= cGaps;
-                            }
-                            else if (leftOverCharPixels <= -cGaps)
-                            {
-                                node.LengthTweak -= cGaps;
-                                leftOverCharPixels += cGaps;
-                            }
-                            else
-                            {
-                                node.LengthTweak += leftOverCharPixels;
-                                leftOverCharPixels = 0;
-                            }
-                        }
-
-                        if ((!contract && node == expandEndNode) || (contract && node == contractEndNode))
-                            break;
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        ///     Checks whether to skip trailing space on line because the next word does not
-        ///     fit.
-        ///     We only check one space - the assumption is that if there is more than one,
-        ///     it is a deliberate attempt to insert spaces.
-        /// </summary>
-        /// <param name="node"></param>
-        /// <param name="lengthSoFar"></param>
-        /// <param name="boundWidth"></param>
-        /// <returns></returns>
-        private bool SkipTrailingSpace(TextNode node, float lengthSoFar, float boundWidth)
-        {
-            if (node.Type == TextNodeType.Space && node.Next != null && node.Next.Type == TextNodeType.Word &&
-                node.ModifiedLength + node.Next.ModifiedLength + lengthSoFar > boundWidth)
-            {
-                return true;
-            }
-
-            return false;
-        }
-
-        /// <summary>
-        ///     Creates node list object associated with the text.
-        /// </summary>
-        /// <param name="text"></param>
-        /// <param name="bounds"></param>
-        /// <returns></returns>
-        public ProcessedText ProcessText(string text, SizeF maxSize, QFontAlignment alignment)
-        {
-            //TODO: bring justify and alignment calculations in here
-
-            maxSize.Width = TransformWidthToViewport(maxSize.Width);
-
-            var nodeList = new TextNodeList(text);
-            nodeList.MeasureNodes(GlFont.FontData, Options);
-
-            //we "crumble" words that are two long so that that can be split up
-            var nodesToCrumble = new List<TextNode>();
-            foreach (TextNode node in nodeList)
-                if ((!Options.WordWrap || node.Length >= maxSize.Width) && node.Type == TextNodeType.Word)
-                    nodesToCrumble.Add(node);
-
-            foreach (TextNode node in nodesToCrumble)
-                nodeList.Crumble(node, 1);
-
-            //need to measure crumbled words
-            nodeList.MeasureNodes(GlFont.FontData, Options);
-
-
-            var processedText = new ProcessedText();
-            processedText.textNodeList = nodeList;
-            processedText.maxSize = maxSize;
-            processedText.alignment = alignment;
-
-
-            return processedText;
-        }
 
         public void Begin()
         {
@@ -1139,16 +331,16 @@ void main(void)
 
         private void InitVBOs()
         {
-            VertexArrayObjects = new QVertexArrayObject[GlFont.FontData.Pages.Length];
+            VertexArrayObjects = new QVertexArrayObject[GlFontDrawingPimitive.GlFont.FontData.Pages.Length];
 
             for (int i = 0; i < VertexArrayObjects.Length; i++)
             {
-                int textureID = GlFont.FontData.Pages[i].GLTexID;
+                int textureID = GlFontDrawingPimitive.GlFont.FontData.Pages[i].GLTexID;
                 VertexArrayObjects[i] = new QVertexArrayObject(QFontSharedState, textureID);
             }
 
-            if (GlFont.FontData.dropShadow != null)
-                GlFont.FontData.dropShadow.InitVBOs();
+            if (GlFontDrawingPimitive.GlFont.FontData.dropShadow != null)
+                GlFontDrawingPimitive.GlFont.FontData.dropShadow.InitVBOs();
         }
 
         public void ResetVBOs()
@@ -1156,8 +348,8 @@ void main(void)
             foreach (QVertexArrayObject buffer in VertexArrayObjects)
                 buffer.Reset();
 
-            if (GlFont.FontData.dropShadow != null)
-                GlFont.FontData.dropShadow.ResetVBOs();
+            if (GlFontDrawingPimitive.GlFont.FontData.dropShadow != null)
+                GlFontDrawingPimitive.GlFont.FontData.dropShadow.ResetVBOs();
         }
 
         public void LoadVBOs()
@@ -1165,8 +357,8 @@ void main(void)
             foreach (QVertexArrayObject buffer in VertexArrayObjects)
                 buffer.Load();
 
-            if (GlFont.FontData.dropShadow != null)
-                GlFont.FontData.dropShadow.LoadVBOs();
+            if (GlFontDrawingPimitive.GlFont.FontData.dropShadow != null)
+                GlFontDrawingPimitive.GlFont.FontData.dropShadow.LoadVBOs();
         }
 
         public void DrawVBOs()
@@ -1175,8 +367,8 @@ void main(void)
             GL.Uniform1(InstanceSharedState.ShaderVariables.SamplerLocation, 0);
             GL.ActiveTexture(InstanceSharedState.DefaultTextureUnit);
 
-            if (GlFont.FontData.dropShadow != null)
-                GlFont.FontData.dropShadow.GlFont.DrawVBOs();
+            if (GlFontDrawingPimitive.GlFont.FontData.dropShadow != null)
+                GlFont.FontData.dropShadow.GlFontDrawingPimitive.GlFont.DrawVBOs();
 
             foreach (QVertexArrayObject buffer in VertexArrayObjects)
                 buffer.Draw();
@@ -1186,7 +378,7 @@ void main(void)
 
         // Track whether Dispose has been called.
         private bool disposed;
-        private readonly GlFont _glFont;
+        private readonly GlFontDrawingPimitive _glFontDrawingPimitive;
 
         /// <summary>
         /// Releases unmanaged and - optionally - managed resources.
@@ -1219,7 +411,7 @@ void main(void)
                 // and unmanaged resources.
                 if (disposing)
                 {
-                    GlFont.FontData.Dispose();
+                    GlFontDrawingPimitive.GlFont.FontData.Dispose();
                     foreach (QVertexArrayObject buffer in VertexArrayObjects)
                         buffer.Dispose();
                 }
