+++ QuickFont/QuickFont.csproj
@@ -51,6 +51,7 @@
     <Compile Include="Configuration\QFontConfiguration.cs" />
     <Compile Include="Configuration\QFontKerningConfiguration.cs" />
     <Compile Include="GlFont.cs" />
+    <Compile Include="GlFontDrawingPimitive.cs" />
     <Compile Include="Helper.cs" />
     <Compile Include="KerningCalculator.cs" />
     <Compile Include="ViewportHelper.cs" />
